Fix LambdaParameterInRestartableEffect conflict with `onDispose` lambda
Composables which accept a parameter named `onDispose` will always raise
a warning for its use in `DisposableEffect`, from the call to
`DisposableEffectScope.onDispose`. This is a rare edge case but doesn't
allow any workarounds other than renaming the lambda or suppressing.

To fix, I've implemented simple plain text matching to ignore any usages
of `onDispose` in a `DisposableEffect`. This may not be perfect (as the
TODO in test cases shows) but in the absence of type resolution it might
be a reasonable workaround. Very open to more elegant solutions!