Fix parameter order handling to allow lambdas with default values before required params in composables (#556)
Allows a bit of leniency for special cases where actually moving a
(non-`@Composable`) lambda to after the required values would make it a
trailing lambda, which wouldn't be desirable either. Basically, going
with the lesser of two evils approach here.

Basically, it makes the ParameterOrder rule be mindful of
https://mrmans0n.github.io/compose-rules/rules/#avoid-trailing-lambdas-for-event-handlers

Fixes #416