Add functional tests  (#559)
I am trying to decide whether to keep the samples/ for tests (they are
mostly for sanity check, but can also be used to showcase working setups
for people) and having a separate set of functional tests could be
reasonable, or a pain in the ass.

We'll see how much maintenance they cause, and see which ones are more
interesting to keep (one or both?).